add windows build support
proxydhcp.go: Removed the platform-dependent enableBroadcast function and the syscall import.

New file proxydhcp_unix.go: Sets socket broadcast option on Unix systems (using int(fd) conversion).

New file proxydhcp_windows.go: Sets socket broadcast option on Windows systems (using syscall.Handle(fd) conversion).

All newly added files use Go 1.17+ //go:build build tags.
