+++ internal/proxydhcp/proxydhcp.go
@@ -11,7 +11,6 @@ import (
 	"net"
 	"strconv"
 	"sync"
-	"syscall"
 
 	"bootimus/internal/metrics"
 
@@ -186,20 +185,6 @@ func pxeVendorOptions() []byte {
 		0x06, 0x01, 0x08, // PXE_DISCOVERY_CONTROL = USE_BOOT_FILE
 		0xff, // end
 	}
-}
-
-func enableBroadcast(conn *net.UDPConn) error {
-	raw, err := conn.SyscallConn()
-	if err != nil {
-		return err
-	}
-	var setErr error
-	if err := raw.Control(func(fd uintptr) {
-		setErr = syscall.SetsockoptInt(int(fd), syscall.SOL_SOCKET, syscall.SO_BROADCAST, 1)
-	}); err != nil {
-		return err
-	}
-	return setErr
 }
 
 func (s *Server) bootfileFor(req *dhcpv4.DHCPv4) string {
