+++ internal/proxydhcp/proxydhcp_windows.go
@@ -1,0 +1,22 @@
+//go:build windows
+
+package proxydhcp
+
+import (
+	"net"
+	"syscall"
+)
+
+func enableBroadcast(conn *net.UDPConn) error {
+	raw, err := conn.SyscallConn()
+	if err != nil {
+		return err
+	}
+	var setErr error
+	if err := raw.Control(func(fd uintptr) {
+		setErr = syscall.SetsockoptInt(syscall.Handle(fd), syscall.SOL_SOCKET, syscall.SO_BROADCAST, 1)
+	}); err != nil {
+		return err
+	}
+	return setErr
+}
